V4L: Add driver for Samsung S5K8B1YX image sensor
This patch adds a stub driver for Samsung S5K8B2YX raw image sensor.
There is currently no I2C communication done in this driver. It can
be added if required. This driver was tested only with Exynos FIMC-IS
subsystem (camera ISP), which communicates with the sensor over I2C
bus in firmware.

Change-Id: I40ea0b1d6bbb139b85f60941343fd8207f4ece8a
Signed-off-by: Sylwester Nawrocki <s.nawrocki@samsung.com>