packaging: move modules from /boot/lib/modules to /boot/modules.img
The latest e2fsprogs have a functionality to make ext2/3/4 image from directory.
It acts like genext2fs or mkext4fs of android tools. I took some patches for it
from e2fsprogs next branch and it will be merged in next release.

A mobile device tend to locate kernel image and modules in the same partition
because it should be easily recovered when the partition was broken. Thus, I
also located the modules image in the /boot partition instead of /lib/modules
directory. This patch supports making the modules.img from obs/gbs build.

Change-Id: I480837378e07bc327ac2379211893799bb452402
Signed-off-by: Chanho Park <chanho61.park@samsung.com>