drm/exynos: dsi: introduce new state for video output availability
Because DSIM_STATE_ENABLED becomes to represent whether DSI can be used
for data transfer or not, this state can be used for checking whether
video ouput is available or not anymore. So new state have to be
introduced. The stete DSIM_STATE_VIDOUT_AVAILABLE represents whether
DSI is prepared for outputting video to a panel or not.

Signed-off-by: Hyungwon Hwang <human.hwang@samsung.com>