drm/exynos: gsc: fix wrong bitwise operation for swap detection
The bits for rotation are not used as exclusively. So GSC_IN_ROT_270 can
not be used for swap detection. The definition of it is same with
GSC_IN_ROT_MASK. It is enough to check GSC_IN_ROT_90 bit is set or not to
check whether width / height size swapping is needed.

Change-Id: I69a98c170b4d31ee68dfd816dfecfd6045d84a44
Signed-off-by: Hyungwon Hwang <human.hwang@samsung.com>