regulator: max77693: Also manipulate the fast charge current
For MAX77693, the fast charge current also needs to be manipulated for
proper charging. The fast charge current is only set in the case of
the MAX77693 type, as the MAX77843 properly manipulates the fast
charge current.
The fast charge current is set to the next possible value below the
maximum input current.

Change-Id: Ieb1664d61ff0b5a44b4c99cd0064d791b9d1c1b6
Signed-off-by: Wolfgang Wiedmeyer <wolfgit@wiedmeyer.de>