PM / AVS: exynos-avs: Support all ASV groups for Exynos 5433 CPU cores
Since multiple supply voltages for a operating frequency and the
'supported-hw' property are not supported by OPPv1, there is no way to
support all the frequency-voltage tables under the dt-based common
device driver. Therefore, the previous version of Exynos5433-specific
device driver only supports the optimized OPP table for the ASV group
number of 8 in version 2 of the OPP table (which means that the driver
supports only one case of all).

However, since there is a requirement to improve the performance under
the given thermal budget for TM2 boards, it is required to support all
frequency-voltage tables corresponding to their ASV groups and table
versions for the Exynos 5433 CPU cores.

This patch modifies the common and Exynos5433 SoC specific AVS device
driver to use static frequency-voltage tables rather than a OPP table
from the DT file.

Change-Id: I74c21403635fc63143d60461b534828fc87cb3f4
Signed-off-by: Wook Song <wook16.song@samsung.com>
Signed-off-by: Chanwoo Choi <cw00.choi@samsung.com>